Edit product link in inspector
